move PA-related impl to evm_adapter.rs
