+++ aarm/src/transaction.rs
@@ -1,11 +1,11 @@
 use crate::action::create_multiple_actions;
-use crate::{action::Action, evm_adapter::AdapterTransaction};
+use crate::action::Action;
 use aarm_core::delta_proof::{DeltaInstance, DeltaProof, DeltaWitness};
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Transaction {
-    pub action: Vec<Action>,
+    pub actions: Vec<Action>,
     pub delta_proof: Delta,
 }
 
@@ -16,9 +16,9 @@ pub enum Delta {
 }
 
 impl Transaction {
-    pub fn new(action: Vec<Action>, delta: Delta) -> Self {
+    pub fn new(actions: Vec<Action>, delta: Delta) -> Self {
         Transaction {
-            action,
+            actions,
             delta_proof: delta,
         }
     }
@@ -42,7 +42,7 @@ impl Transaction {
                 if DeltaProof::verify(&msg, proof, instance).is_err() {
                     return false;
                 }
-                for action in &self.action {
+                for action in &self.actions {
                     if !action.verify() {
                         return false;
                     }
@@ -55,7 +55,7 @@ impl Transaction {
 
     pub fn get_delta_instance(&self) -> DeltaInstance {
         let deltas = self
-            .action
+            .actions
             .iter()
             .flat_map(|action| action.get_delta())
             .collect::<Vec<_>>();
@@ -64,34 +64,22 @@ impl Transaction {
 
     pub fn get_delta_msg(&self) -> Vec<u8> {
         let mut msg = Vec::new();
-        for action in &self.action {
+        for action in &self.actions {
             msg.extend(action.get_delta_msg());
         }
         msg
     }
 
     pub fn compose(tx1: Transaction, tx2: Transaction) -> Transaction {
-        let mut action = tx1.action;
-        action.extend(tx2.action);
+        let mut actions = tx1.actions;
+        actions.extend(tx2.actions);
         let delta = match (&tx1.delta_proof, &tx2.delta_proof) {
             (Delta::Witness(witness1), Delta::Witness(witness2)) => {
                 Delta::Witness(witness1.compose(witness2))
             }
             _ => panic!("Cannot compose transactions with different delta types"),
         };
-        Transaction::new(action, delta)
-    }
-
-    pub fn convert(&self) -> AdapterTransaction {
-        let actions = self.action.iter().map(|action| action.convert()).collect();
-        let delta_proof = match &self.delta_proof {
-            Delta::Witness(_) => panic!("Unbalanced Transactions cannot be converted"),
-            Delta::Proof(proof) => proof.to_bytes().to_vec(),
-        };
-        AdapterTransaction {
-            actions,
-            delta_proof,
-        }
+        Transaction::new(actions, delta)
     }
 }
 
@@ -99,8 +87,7 @@ pub fn generate_test_transaction(n_actions: usize) -> Transaction {
     let (actions, delta_witness) = create_multiple_actions(n_actions);
     let mut tx = Transaction::new(actions, Delta::Witness(delta_witness));
     tx.generate_delta_proof();
-    assert!(tx.verify()); // TODO move into separate test
-    let _adapter_tx = tx.convert();
+    assert!(tx.verify());
     tx
 }
